refactored purchase structure
