+++ drizzle/0037_fast_spacker_dave.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "users" RENAME COLUMN "rate" TO "quota";
