+++ src/lib/auth/lucia.ts
@@ -23,7 +23,7 @@ export const lucia = new Lucia(adapter, {
 			paid: attributes.paid,
 			stripeId: attributes.stripeId,
 			joined: attributes.joined,
-			rate: attributes.rate,
+			quota: attributes.quota,
 			emailVerified: attributes.emailVerified
 		};
 	}
@@ -44,6 +44,6 @@ interface DatabaseUserAttributes {
 	paid: boolean;
 	stripeId: string | null;
 	joined: Date;
-	rate: number;
+	quota: number;
 	emailVerified: boolean;
 }
