+++ src/routes/+page.svelte
@@ -28,7 +28,7 @@
 		Hi, <span class="font-bold text-primary">{user.username}</span>
 	</h3>
 	<h1 class="heading-main text-center">What are your plans?</h1>
-	{#if user.paid}
+	{#if user.quota < 3 || user.paid}
 		<form action="?/create" method="POST" class="my-12 flex flex-col" use:enhance>
 			<div class="flex items-center justify-center gap-4">
 				<input
@@ -46,20 +46,18 @@
 					{/if}
 				</button>
 			</div>
+			{#if !user.paid}
+				<p class="text-muted mt-2">Free remaining: {3 - user.quota}</p>
+			{/if}
 			{#if rateLimited}
 				<span class="mt-2 text-error">Too many requests. Try again later</span>
 			{/if}
 		</form>
 	{:else}
-		<div class="my-6">
-			<form action="/?/purchase" method="post">
-				<button
-					class="btn btn-warning tooltip tooltip-top"
-					data-tip="Purchase the product to continue"
-					>Purchase ($20)
-				</button>
-			</form>
-		</div>
+		<p class="my-6 text-warning">
+			You have used all of your free creations. <a href="/account/billing">purchase</a> the product to
+			continue.
+		</p>
 	{/if}
 
 	<section class="mb-8 flex w-full max-w-2xl flex-col items-center gap-4">
