+++ src/routes/account/billing/+page.svelte
@@ -41,7 +41,7 @@
 		</ul>
 	{:else}
 		<p class="mb-2 text-warning">Please purchase the product to use the features of this app.</p>
-		<form action="/?/purchase" method="post">
+		<form action="?/purchase" method="post">
 			<button class="btn btn-warning">Purchase ($20)</button>
 		</form>
 	{/if}
